requirements.txt
argon2-cffi==19.2.0
cffi==1.13.1
Click==7.0
dnspython==1.16.0
eventlet==0.25.1
ffmpeg-python==0.2.0
Flask==1.1.1
future==0.17.1
greenlet==0.4.15
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
monotonic==1.5
mypy==0.730
mypy-extensions==0.4.2
passlib==1.7.1
pkg-resources==0.0.0
pycparser==2.19
python-dotenv==0.10.3
six==1.12.0
SQLAlchemy==1.3.10
tinytag==1.2.2
typed-ast==1.4.0
typing-extensions==3.7.4
Werkzeug==0.16.0
